8151413: os::allocation_granularity/page_size and friends return signed values

+++ src/hotspot/cpu/aarch64/interp_masm_aarch64.cpp
@@ -802,7 +802,7 @@ void InterpreterMacroAssembler::lock_object(Register lock_reg)
     // copy
     mov(rscratch1, sp);
     sub(swap_reg, swap_reg, rscratch1);
-    ands(swap_reg, swap_reg, (uint64_t)(7 - os::vm_page_size()));
+    ands(swap_reg, swap_reg, (uint64_t)(7 - (int)os::vm_page_size()));
 
     // Save the test result, for recursive case, the result is zero
     str(swap_reg, Address(lock_reg, mark_offset));
